+++ horde_client/client.py
@@ -98,7 +98,7 @@ class HordeClient:
         return response.json()
 
 
-    def list_models(self, type:model.ModelType) -> List[model.Model]:
+    def list_models(self, type:model.ModelType, sort_by:model.SortBy = model.SortBy.perf) -> List[model.Model]:
         '''
         '''
         results = self.get(
@@ -112,6 +112,21 @@ class HordeClient:
             model_list.append(
                 model.Model(**model_obj)
             )
+
+        # Sort model results
+        sort_by_fn = None
+        reverse = False
+
+        if sort_by == model.SortBy.perf:
+            reverse = True
+            sort_by_fn = lambda model: model.performance
+        
+        elif sort_by == model.SortBy.name:
+            reverse = False
+            sort_by_fn = lambda model: model.name
+
+        model_list = sorted(model_list, key=sort_by_fn, reverse=reverse)
+
         return model_list
     
     def clear_model(self):
@@ -124,11 +139,22 @@ class HordeClient:
         '''
         self.__models.append(model)
     
-    def check_job_status(self, job:model.Job) -> model.JobResponse:
+    def check_job_status(self, job:model.Job, type:model.ModelType, progress=False) -> model.JobResponse:
         '''
+        Get latest status of Horde Async Job
         '''
+        url = None
+
+        if type == model.ModelType.text:
+            url = config.ENDPOINT_LIST['V2__ASYNC_TEXT_STATUS']
+        
+        elif progress == True and type == model.ModelType.image:
+            url = config.ENDPOINT_LIST['V2__ASYNC_IMAGE_PROGRESS_STATUS']
+        elif progress == False and type == model.ModelType.image:
+            url = config.ENDPOINT_LIST['V2__ASYNC_IMAGE_STATUS']
+
         status = self.get(
-            config.ENDPOINT_LIST['V2__ASYNC_TEXT_STATUS'],
+            url,
             path_params={
                 'id': job.id
             }
@@ -137,6 +163,7 @@ class HordeClient:
 
     def text_gen(self, prompt:str, params:model.TextGenParams = model.TextGenParams()) -> model.JobResponse:
         '''
+        Leverage Horde LLM service to complete initial prompt using text generation.
         '''
         request_data = model.TextGenRequest(
             models = self.__models,
@@ -154,10 +181,36 @@ class HordeClient:
         )
 
         while True:
-            job_status = self.check_job_status(job)
+            job_status = self.check_job_status(job, model.ModelType.text)
             if job_status.done or not job_status.is_possible:
                 return job_status
             time.sleep(config.REQUEST_RETRY_TIMEOUT)
         
+    def image_gen(self, prompt: str, params:model.ImageGenParams = model.ImageGenParams()) -> model.JobResponse:
+        '''
+        Leverage Horde LLM service to generate image using an initial prompt.
+        '''
+        request_data = model.ImageGenRequest(
+            models = self.__models,
+            params = params,
+            prompt = prompt
+        )
 
+        job_result = self.post(
+            config.ENDPOINT_LIST['V2__ASYNC_IMAGE_SUBMIT'],
+            json_payload = request_data.model_dump()
+        )
+
+        job = model.Job(
+            **job_result
+        )
+
+        while True:
+            job_status = self.check_job_status(job, model.ModelType.image, progress=True)
+
+            if job_status.done or not job_status.is_possible:
+                break
+            time.sleep(config.REQUEST_RETRY_TIMEOUT)
+
+        return self.check_job_status(job, model.ModelType.image)