+++ horde_client/model.py
@@ -2,6 +2,10 @@ from enum import Enum
 from pydantic import BaseModel
 from typing import List
 
+class SortBy(str, Enum):
+    name = 'name'
+    perf = 'perf'
+
 class ModelType(str, Enum):
     text = 'text'
     image = 'image'
@@ -20,6 +24,10 @@ class Model(BaseModel):
             name = self.name,
             perf = self.performance
         )
+
+'''
+Text Generation
+'''
 
 class TextGenParams(BaseModel):
     frmtadsnsp: bool = False
@@ -40,7 +48,6 @@ class TextGenParams(BaseModel):
     top_p: float = 0.92
     typical: int = 1
 
-
 class TextGenRequest(BaseModel):
     models: List[str]
     params: TextGenParams
@@ -59,15 +66,22 @@ class JobGenerationOutputState(str, Enum):
 class JobGenerationOutput(BaseModel):
     worker_id: str
     worker_name: str
-    text: str
     state: JobGenerationOutputState
     seed: int
     model: str
 
+    # Text Generation Parameters
+    text: str | None = None
+
+    # Image Generation Parameters
+    id: str | None = None
+    img: str | None = None
+
+
 class JobResponse(BaseModel):
     done: bool
     faulted: bool
-    generations: List[JobGenerationOutput]
+    generations: List[JobGenerationOutput | None] = []
     is_possible: bool
     kudos: int
     processing: int
@@ -77,8 +91,50 @@ class JobResponse(BaseModel):
     waiting: int
 
     def __str__(self) -> str:
-        return "Output(done={done},responses=[{responses}])".format(
+        return "Output(done={done},responses=[{responses}], queue_position={queue_position})".format(
             done = str(self.done),
-            responses = str(len(self.generations))
+            responses = str(len(self.generations)),
+            queue_position = str(self.queue_position)
         )
 
+
+'''
+Image Generation
+'''
+
+class ImageGenSampleType(str, Enum):
+    k_lms = "k_lms"
+    k_euler_a = "k_euler_a"
+
+class ImageGenLora(BaseModel):
+    name: str
+    model: int = 1
+    clip: int = 1
+    inject_trigger: str = "string"
+
+class ImageGenParams(BaseModel):
+    width: int = 512
+    height: int = 512
+    
+    n: int = 1
+    steps: int = 20
+    
+    seed: str = ""
+    sampler_name: str = ImageGenSampleType.k_euler_a
+    karras: bool = False
+    cfg_scale: float = 7.0
+    post_processing: List[str | None] = []
+    loras: List[ImageGenLora | None] = []
+
+
+class ImageGenRequest(BaseModel):
+    models: List[str]
+    params: ImageGenParams
+    prompt: str
+    use_default_badwordsids: bool = True
+    workers: List[str] = []
+
+    censor_nsfw: bool = True
+    nsfw: bool = False
+    r2: bool = False
+    replacement_filter: bool = True